requirements.txt
census==0.8.20
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
setuptools==69.0.2
six==1.16.0
urllib3==2.1.0
